NTBS-3078: Split program configuration into separate files
